+++ changelog.md
@@ -1,7 +1,7 @@
 # Version 2023.1.5 (2023-01-17)
 - Remove LOWBAT from HM-LC-Sw1-DR
 - Sort lists in parameter_visibility.py
-- 
+- Replace custom entity config data structure by CustomConfig
 
 # Version 2023.1.4 (2023-01-16)
 - Remove obsolete parse_ccu_sys_var